# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/cofre/unlock": {
      "post": {
        "summary": "Unlock cofre with PIN",
        "tags": [
          "Cofre"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pin": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "pin"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "423": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/cofre/entries": {
      "get": {
        "summary": "List cofre entries",
        "tags": [
          "Cofre"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "credential",
                "secure_note",
                "api_key"
              ]
            },
            "in": "query",
            "name": "category",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string",
                            "enum": [
                              "credential",
                              "secure_note",
                              "api_key"
                            ]
                          },
                          "title": {
                            "type": "string"
                          },
                          "url": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "username": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "provider": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "category",
                          "title",
                          "url",
                          "username",
                          "provider",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "entries"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create cofre entry",
        "tags": [
          "Cofre"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string",
                    "enum": [
                      "credential",
                      "secure_note",
                      "api_key"
                    ]
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "url": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "username": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "password": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "content": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "provider": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "token": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  }
                },
                "required": [
                  "category",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entry": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string",
                          "enum": [
                            "credential",
                            "secure_note",
                            "api_key"
                          ]
                        },
                        "title": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "category",
                        "title",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "entry"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}